Finished refactor of file readin loop to make it a stateful entity rather than a mess of references being passed around
